+++ to-gather-frontend/src/components/Footer.css
