+++ to-gather-frontend/src/components/Footer.jsx
@@ -1,13 +1,10 @@
 import React from "react";
-//import InstagramIcon from "@material-ui/icons/Instagram";
-//import TwitterIcon from "@material-ui/icons/Twitter";
-//import FacebookIcon from "@material-ui/icons/Facebook";
-//import LinkedInIcon from "@material-ui/icons/LinkedIn";
+//npm install @mui/icons-material
 import LinkedInIcon from "@mui/icons-material/LinkedIn";
 import InstagramIcon from "@mui/icons-material/Instagram";
 import TwitterIcon from "@mui/icons-material/Twitter";
 import FacebookIcon from "@mui/icons-material/Facebook";
-import "../styles/Footer.css";
+import "./Footer.css";
 
 function Footer() {
   return (
